Code re-org + fixed Router
